Split Method into separate files
